Revert "Move fixture TOML coverage to Vitest"
This reverts commit 4f068ec37e397955a63c91f974e7d5280a07cf05.
