+++ packages/e2e/playwright.config.ts
@@ -26,7 +26,7 @@ export default defineConfig({
   projects: [
     {
       name: 'local',
-      testMatch: ['tests/smoke.spec.ts', 'tests/smoke-pty.spec.ts'],
+      testMatch: ['tests/smoke.spec.ts', 'tests/smoke-pty.spec.ts', 'tests/fixture-toml.spec.ts'],
     },
     {
       name: 'remote-auth',
@@ -35,7 +35,7 @@ export default defineConfig({
     {
       name: 'remote',
       testMatch: 'tests/*.spec.ts',
-      testIgnore: ['tests/smoke.spec.ts', 'tests/smoke-pty.spec.ts'],
+      testIgnore: ['tests/smoke.spec.ts', 'tests/smoke-pty.spec.ts', 'tests/fixture-toml.spec.ts'],
       dependencies: ['remote-auth'],
     },
   ],
