+++ packages/e2e/setup/app.ts
@@ -2,7 +2,6 @@
 import {authFixture} from './auth.js'
 import {getLastPageStatus, isVisibleWithin, navigateToDashboard, refreshIfPageError} from './browser.js'
 import {CLI_TIMEOUT, BROWSER_TIMEOUT} from './constants.js'
-import {mergeFixtureToml} from '../helpers/fixture-toml.js'
 import * as toml from '@iarna/toml'
 import * as path from 'path'
 import * as fs from 'fs'
@@ -116,6 +115,38 @@ export function extractClientId(appDir: string, configName?: string): string {
  */
 export function devDashboardAppUrl(appDir: string, orgId: string, configName?: string): string {
   return `https://dev.shopify.com/dashboard/${orgId}/apps/${extractClientId(appDir, configName)}`
+}
+
+/**
+ * Merge an E2E TOML fixture into the app config generated by the template.
+ * Template-owned fields not mentioned by the fixture are preserved.
+ */
+type TomlTable = Parameters<typeof toml.stringify>[0]
+
+function isTomlTable(value: unknown): value is TomlTable {
+  return typeof value === 'object' && value !== null && !Array.isArray(value) && !(value instanceof Date)
+}
+
+function mergeTomlTables(base: TomlTable, overlay: TomlTable): TomlTable {
+  const merged: TomlTable = {...base}
+  for (const [key, overlayValue] of Object.entries(overlay)) {
+    const baseValue = merged[key]
+    merged[key] =
+      isTomlTable(baseValue) && isTomlTable(overlayValue) ? mergeTomlTables(baseValue, overlayValue) : overlayValue
+  }
+  return merged
+}
+
+export function mergeFixtureToml(generatedTomlContent: string, fixtureTomlContent: string, name: string): string {
+  const generated = toml.parse(generatedTomlContent)
+  const clientId = generated.client_id as string | undefined
+  if (!clientId) {
+    throw new Error('Could not find client_id in generated shopify.app.toml')
+  }
+
+  const fixture = toml.parse(fixtureTomlContent)
+
+  return toml.stringify(mergeTomlTables(generated, {...fixture, client_id: clientId, name}))
 }
 
 export function injectFixtureToml(appDir: string, fixtureTomlContent: string, name: string): void {
