+++ vite.config.ts
@@ -12,7 +12,6 @@ export default defineConfig({
       'packages/app/vite.config.ts',
       'packages/cli/vite.config.ts',
       'packages/cli-kit/vite.config.ts',
-      'packages/e2e/vite.config.mts',
       'packages/eslint-plugin-cli/vite.config.ts',
       'packages/organizations/vite.config.ts',
       'packages/plugin-cloudflare/vite.config.ts',
